% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doer.R
\name{install_matplotlib}
\alias{install_matplotlib}
\title{install python library matplotlib}
\usage{
install_matplotlib(method = "auto", conda = "auto")
}
\arguments{
\item{method}{PARAM_DESCRIPTION, Default: 'auto'}

\item{conda}{PARAM_DESCRIPTION, Default: 'auto'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
install_matplotlib()
 }
}
}
\seealso{
\code{\link[reticulate]{py_install}}
}
